refactor: useMap 훅으로 분리
