+++ src/hooks/index.ts
@@ -1,3 +1,3 @@
+export * from "./useMap";
 export * from "./useTimeoutFn";
 export * from "./useTimeout";
-
